# packages/server/dist/swagger.json
{
	"basePath": "/",
	"definitions": {
...
		"KeyValue": {
			"description": "",
			"properties": {
				"label": {
					"type": "string",
					"description": ""
				},
				"value": {
					"type": "string",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"label",
				"value"
			]
		},
		"Error": {
			"description": "",
			"properties": {
				"name": {
					"type": "string",
					"description": ""
				},
				"message": {
					"type": "string",
					"description": ""
				},
				"stack": {
					"type": "string",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"name",
				"message"
			]
		},
...
		"MenuResponse": {
			"description": "",
			"properties": {
				"id": {
					"type": "string",
					"description": ""
				},
				"name": {
					"type": "string",
					"description": ""
				},
				"slug": {
					"type": "string",
					"description": ""
				},
				"group": {
					"type": "boolean",
					"description": ""
				},
				"link": {
					"type": "string",
					"description": ""
				},
				"externalLink": {
					"type": "string",
					"description": ""
				},
				"blank": {
					"type": "boolean",
					"description": ""
				},
				"icon": {
					"type": "string",
					"description": ""
				},
				"order": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"enable": {
					"type": "boolean",
					"description": ""
				},
				"expanded": {
					"type": "boolean",
					"description": ""
				},
				"acl": {
					"type": "string",
					"description": ""
				},
				"permissions": {
					"type": "array",
					"items": {
						"type": "string"
					},
					"description": ""
				},
				"parent": {
					"type": "object",
					"description": ""
				},
				"isMenu": {
					"type": "boolean",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"id",
				"name",
				"slug",
				"group",
				"link",
				"externalLink",
				"blank",
				"icon",
				"order",
				"enable",
				"expanded",
				"acl",
				"isMenu"
			]
		},
		"CreateMenuDto": {
			"description": "",
			"properties": {
				"name": {
					"type": "string",
					"description": ""
				},
				"slug": {
					"type": "string",
					"description": ""
				},
				"group": {
					"type": "boolean",
					"description": ""
				},
				"link": {
					"type": "string",
					"description": ""
				},
				"externalLink": {
					"type": "string",
					"description": ""
				},
				"blank": {
					"type": "boolean",
					"description": ""
				},
				"icon": {
					"type": "string",
					"description": ""
				},
				"order": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"enable": {
					"type": "boolean",
					"description": ""
				},
				"expanded": {
					"type": "boolean",
					"description": ""
				},
				"acl": {
					"type": "string",
					"description": ""
				},
				"permissions": {
					"type": "array",
					"items": {
						"type": "string"
					},
					"description": ""
				},
				"parent": {
					"type": "object",
					"description": ""
				},
				"isMenu": {
					"type": "boolean",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"name",
				"slug",
				"group",
				"link",
				"externalLink",
				"blank",
				"icon",
				"order",
				"enable",
				"expanded",
				"acl",
				"isMenu"
			]
		},
		"EditMenuDto": {
			"description": "",
			"properties": {
				"id": {
					"type": "string",
					"description": ""
				},
				"name": {
					"type": "string",
					"description": ""
				},
				"slug": {
					"type": "string",
					"description": ""
				},
				"group": {
					"type": "boolean",
					"description": ""
				},
				"link": {
					"type": "string",
					"description": ""
				},
				"externalLink": {
					"type": "string",
					"description": ""
				},
				"blank": {
					"type": "boolean",
					"description": ""
				},
				"icon": {
					"type": "string",
					"description": ""
				},
				"order": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"enable": {
					"type": "boolean",
					"description": ""
				},
				"expanded": {
					"type": "boolean",
					"description": ""
				},
				"acl": {
					"type": "string",
					"description": ""
				},
				"permissions": {
					"type": "array",
					"items": {
						"type": "string"
					},
					"description": ""
				},
				"parent": {
					"type": "object",
					"description": ""
				},
				"isMenu": {
					"type": "boolean",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"id",
				"name",
				"slug",
				"group",
				"link",
				"externalLink",
				"blank",
				"icon",
				"order",
				"enable",
				"expanded",
				"acl",
				"isMenu"
			]
		},
		"PaginateMenu": {
			"description": "",
			"properties": {
				"error": {
					"$ref": "#/definitions/Error"
				},
				"list": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/MenuResponse"
					},
					"description": ""
				},
				"total": {
					"type": "number",
					"format": "double",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"list",
				"total"
			]
		},
		"SettingResponse": {
			"description": "",
			"properties": {
				"id": {
					"type": "string",
					"description": ""
				},
				"name": {
					"type": "string",
					"description": ""
				},
				"key": {
					"type": "string",
					"description": ""
				},
				"value": {
					"type": "object",
					"description": ""
				},
				"description": {
					"type": "string",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"id",
				"name",
				"key",
				"value",
				"description"
			]
		},
		"CreateSettingDto": {
			"description": "",
			"properties": {
				"name": {
					"type": "string",
					"description": ""
				},
				"key": {
					"type": "string",
					"description": ""
				},
				"value": {
					"type": "object",
					"description": ""
				},
				"description": {
					"type": "string",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"name",
				"key",
				"value",
				"description"
			]
		},
		"EditSettingDto": {
			"description": "",
			"properties": {
				"id": {
					"type": "string",
					"description": ""
				},
				"name": {
					"type": "string",
					"description": ""
				},
				"key": {
					"type": "string",
					"description": ""
				},
				"value": {
					"type": "object",
					"description": ""
				},
				"description": {
					"type": "string",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"id",
				"name",
				"key",
				"value",
				"description"
			]
		},
		"PaginateSetting": {
			"description": "",
			"properties": {
				"error": {
					"$ref": "#/definitions/Error"
				},
				"list": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/SettingResponse"
					},
					"description": ""
				},
				"total": {
					"type": "number",
					"format": "double",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"list",
				"total"
			]
		},
...
		"RoleResponse": {
			"description": "",
			"properties": {
				"id": {
					"type": "string",
					"description": ""
				},
				"name": {
					"type": "string",
					"description": ""
				},
				"description": {
					"type": "string",
					"description": ""
				},
				"permissions": {
					"type": "array",
					"items": {
						"type": "string"
					},
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"id",
				"name",
				"description",
				"permissions"
			]
		},
		"CreateRoleDto": {
			"description": "",
			"properties": {
				"name": {
					"type": "string",
					"description": ""
				},
				"description": {
					"type": "string",
					"description": ""
				},
				"permissions": {
					"type": "array",
					"items": {
						"type": "string"
					},
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"name",
				"description",
				"permissions"
			]
		},
		"EditRoleDto": {
			"description": "",
			"properties": {
				"id": {
					"type": "string",
					"description": ""
				},
				"name": {
					"type": "string",
					"description": ""
				},
				"description": {
					"type": "string",
					"description": ""
				},
				"permissions": {
					"type": "array",
					"items": {
						"type": "string"
					},
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"id",
				"name",
				"description",
				"permissions"
			]
		},
		"PaginateRole": {
			"description": "",
			"properties": {
				"error": {
					"$ref": "#/definitions/Error"
				},
				"list": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/RoleResponse"
					},
					"description": ""
				},
				"total": {
					"type": "number",
					"format": "double",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"list",
				"total"
			]
		},
...
	},
	"info": {
		"description": "项目标准接口",
		"license": {
			"name": "MIT"
		},
		"title": "rest api interface",
		"version": "0.0.1"
	},
	"paths": {
...
		"/api/menu/search": {
			"get": {
				"operationId": "MenuSearch",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/KeyValue"
							}
						}
					}
				},
				"description": "查询菜单",
				"tags": [
					"core"
				],
				"parameters": [
					{
						"description": "关键词",
						"in": "query",
						"name": "keyword",
						"required": false,
						"type": "string"
					},
					{
						"description": "已选中的菜单编号",
						"in": "query",
						"name": "value",
						"required": false,
						"type": "string"
					}
				]
			}
		},
		"/api/menu": {
			"post": {
				"operationId": "MenuCreate",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/MenuResponse"
						}
					}
				},
				"description": "创建菜单",
				"tags": [
					"core"
				],
				"consumes": [
					"application/json"
				],
				"parameters": [
					{
						"description": "创建参数",
						"in": "body",
						"name": "entry",
						"required": true,
						"schema": {
							"$ref": "#/definitions/CreateMenuDto"
						}
					}
				]
			},
			"put": {
				"operationId": "MenuUpdate",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/MenuResponse"
						}
					}
				},
				"description": "更新菜单",
				"tags": [
					"core"
				],
				"consumes": [
					"application/json"
				],
				"parameters": [
					{
						"description": "菜单参数",
						"in": "body",
						"name": "entry",
						"required": true,
						"schema": {
							"$ref": "#/definitions/EditMenuDto"
						}
					}
				]
			}
		},
		"/api/menu/query": {
			"get": {
				"operationId": "MenuQuery",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/PaginateMenu"
						}
					}
				},
				"description": "查询菜单数据",
				"tags": [
					"core"
				],
				"parameters": [
					{
						"description": "关键词",
						"in": "query",
						"name": "keyword",
						"required": false,
						"type": "string"
					},
					{
						"description": "",
						"in": "query",
						"name": "isMenu",
						"required": false,
						"type": "boolean"
					},
					{
						"description": "第几页",
						"in": "query",
						"name": "page",
						"required": false,
						"type": "number",
						"format": "double"
					},
					{
						"description": "页大小",
						"in": "query",
						"name": "size",
						"required": false,
						"type": "number",
						"format": "double"
					},
					{
						"description": "排序",
						"in": "query",
						"name": "sort",
						"required": false,
						"type": "string"
					}
				]
			}
		},
		"/api/menu/{id}": {
			"delete": {
				"operationId": "MenuRemove",
				"produces": [
					"text/html"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "boolean"
						}
					}
				},
				"description": "删除菜单",
				"tags": [
					"core"
				],
				"parameters": [
					{
						"description": "菜单编号",
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			},
			"get": {
				"operationId": "MenuGet",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/MenuResponse"
						}
					}
				},
				"description": "查询菜单",
				"tags": [
					"core"
				],
				"parameters": [
					{
						"description": "菜单编号",
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			}
		},
...
		"/api/setting/main": {
			"get": {
				"operationId": "SettingGetMainSettings",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/SettingResponse"
							}
						}
					}
				},
				"description": "获取设置项",
				"tags": [
					"core"
				],
				"parameters": [
					{
						"description": "设置项key的集合",
						"in": "query",
						"name": "keys",
						"required": false,
						"type": "string"
					}
				]
			}
		},
		"/api/setting/key/{name}": {
			"get": {
				"operationId": "SettingGetSettingsByKey",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/SettingResponse"
						}
					}
				},
				"description": "通过Key获取设置项目",
				"tags": [
					"core"
				],
				"parameters": [
					{
						"description": "键名",
						"in": "path",
						"name": "name",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/api/setting/search": {
			"get": {
				"operationId": "SettingSearch",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/SettingResponse"
							}
						}
					}
				},
				"description": "查询设置项",
				"tags": [
					"core"
				],
				"parameters": [
					{
						"description": "关键词",
						"in": "query",
						"name": "keyword",
						"required": false,
						"type": "string"
					},
					{
						"description": "键",
						"in": "query",
						"name": "value",
						"required": false,
						"type": "string"
					}
				]
			}
		},
		"/api/setting": {
			"post": {
				"operationId": "SettingCreate",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/SettingResponse"
						}
					}
				},
				"description": "创建设置项",
				"tags": [
					"core"
				],
				"consumes": [
					"application/json"
				],
				"parameters": [
					{
						"description": "设置项实体",
						"in": "body",
						"name": "entry",
						"required": true,
						"schema": {
							"$ref": "#/definitions/CreateSettingDto"
						}
					}
				]
			},
			"put": {
				"operationId": "SettingUpdate",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/SettingResponse"
						}
					}
				},
				"description": "更新设置项",
				"tags": [
					"core"
				],
				"consumes": [
					"application/json"
				],
				"parameters": [
					{
						"description": "设置项实体",
						"in": "body",
						"name": "entry",
						"required": true,
						"schema": {
							"$ref": "#/definitions/EditSettingDto"
						}
					}
				]
			}
		},
		"/api/setting/query": {
			"get": {
				"operationId": "SettingQuery",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/PaginateSetting"
						}
					}
				},
				"description": "分页查询设置项",
				"tags": [
					"core"
				],
				"parameters": [
					{
						"description": "关键词",
						"in": "query",
						"name": "keyword",
						"required": false,
						"type": "string"
					},
					{
						"description": "第几页",
						"in": "query",
						"name": "page",
						"required": false,
						"type": "number",
						"format": "double"
					},
					{
						"description": "页大小",
						"in": "query",
						"name": "size",
						"required": false,
						"type": "number",
						"format": "double"
					},
					{
						"description": "排序",
						"in": "query",
						"name": "sort",
						"required": false,
						"type": "string"
					}
				]
			}
		},
		"/api/setting/{id}": {
			"get": {
				"operationId": "SettingGet",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/SettingResponse"
						}
					}
				},
				"description": "按编号获取设置项",
				"tags": [
					"core"
				],
				"parameters": [
					{
						"description": "键",
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			},
			"delete": {
				"operationId": "SettingRemove",
				"produces": [
					"text/html"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "boolean"
						}
					}
				},
				"description": "删除设置项",
				"tags": [
					"core"
				],
				"parameters": [
					{
						"description": "键",
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			}
		},
...
		"/api/role/search": {
			"get": {
				"operationId": "RoleSearch",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/KeyValue"
							}
						}
					}
				},
				"description": "搜索角色",
				"tags": [
					"core"
				],
				"parameters": [
					{
						"description": "关键词",
						"in": "query",
						"name": "keyword",
						"required": false,
						"type": "string"
					},
					{
						"description": "键",
						"in": "query",
						"name": "value",
						"required": false,
						"type": "string"
					}
				]
			}
		},
		"/api/role": {
			"post": {
				"operationId": "RoleCreate",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/RoleResponse"
						}
					}
				},
				"description": "创建角色",
				"tags": [
					"core"
				],
				"consumes": [
					"application/json"
				],
				"parameters": [
					{
						"description": "设置项实体",
						"in": "body",
						"name": "entry",
						"required": true,
						"schema": {
							"$ref": "#/definitions/CreateRoleDto"
						}
					}
				]
			},
			"put": {
				"operationId": "RoleUpdate",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/RoleResponse"
						}
					}
				},
				"description": "更新角色",
				"tags": [
					"core"
				],
				"consumes": [
					"application/json"
				],
				"parameters": [
					{
						"description": "设置项实体",
						"in": "body",
						"name": "entry",
						"required": true,
						"schema": {
							"$ref": "#/definitions/EditRoleDto"
						}
					}
				]
			}
		},
		"/api/role/query": {
			"get": {
				"operationId": "RoleQuery",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/PaginateRole"
						}
					}
				},
				"description": "分页查询角色",
				"tags": [
					"core"
				],
				"parameters": [
					{
						"description": "关键词",
						"in": "query",
						"name": "keyword",
						"required": false,
						"type": "string"
					},
					{
						"description": "第几页",
						"in": "query",
						"name": "page",
						"required": false,
						"type": "number",
						"format": "double"
					},
					{
						"description": "页大小",
						"in": "query",
						"name": "size",
						"required": false,
						"type": "number",
						"format": "double"
					},
					{
						"description": "排序",
						"in": "query",
						"name": "sort",
						"required": false,
						"type": "string"
					}
				]
			}
		},
		"/api/role/{id}": {
			"get": {
				"operationId": "RoleGet",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/RoleResponse"
						}
					}
				},
				"description": "按编号获取角色",
				"tags": [
					"core"
				],
				"parameters": [
					{
						"description": "键",
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			},
			"delete": {
				"operationId": "RoleRemove",
				"produces": [
					"text/html"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "boolean"
						}
					}
				},
				"description": "删除角色",
				"tags": [
					"core"
				],
				"parameters": [
					{
						"description": "键",
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			}
		},
...
	},
	"swagger": "2.0",
	"securityDefinitions": {},
	"produces": [
		"application/json"
	]
}